requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
pymupdf==1.24.11
python-docx==1.1.2
python-multipart==0.0.12
sqlite3-fts5-fix==0.0.1; sys_platform == 'linux'
